411: Reformat + rename + clean
