+++ quad_pyblish_module/plugins/photoshop/publish/help/validate_names_unique.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <root>
 <error id="main">
-<title>Name Uniqueness</title>
+<title>Unique Layer Name Validator</title>
 <description>
-## Name of layer not unique
+## Multiple element names aren't unique
 
-Layer or Group have not unique names
+Layers or Groups share the same name, this is not valid.
 
 ### How to repair?
 
-You can select the default layers with the "Select Layers" button on the right and correct them.
+You can select the detected layers with the "Select Layers" button on the right, then you need to change the names and make them all unique.
 </description>
 </error>
 </root>
