+++ quad_pyblish_module/plugins/photoshop/publish/validate_blendmode.py
@@ -39,8 +39,7 @@ class ValidateBlendMode(
     active = False
 
     def process(self, context):
-
-        if not self.is_active(context.data):
+        if not self.active or not self.is_active(context.data):
             return
         
         PASSTHROUGH = "passThrough"
