+++ quad_pyblish_module/plugins/photoshop/publish/validate_layers_organization.py
@@ -82,7 +82,7 @@ class ValidateLayersOrganization(
     active = False
 
     def process(self, context):
-        if not self.is_active(context.data):
+        if not self.active or not self.is_active(context.data):
             return
         
         project_name = os.environ['AVALON_PROJECT']
